Add Reuters superclass and rename ReutersGraph and ReutersText
