+++ data_prep/reuters_text.py
@@ -1,22 +1,16 @@
 #!/usr/bin/env python3
-import random
-from collections import defaultdict
-
 import nltk
-
 nltk.download('reuters')
 from nltk.corpus import reuters
 
 import torch
-from data_prep.dataset import TextDataset
-from torch.utils.data import Dataset
+from data_prep.dataset import TextDataset, Reuters
 
-class Reuters(TextDataset):
+class ReutersText(Reuters, TextDataset):
     def __init__(self, encodings, labels, classes):
         self.encodings = encodings
         self.labels = labels
         self.classes = classes
-
 
     @classmethod
     def splits(cls, tokenizer, r8=False, val_size=0.1):
@@ -48,56 +42,6 @@ class Reuters(TextDataset):
         return cls(encodings, labels, unique_cls)
 
     @staticmethod
-    def prepare_reuters(r8=False, val_size=0.1):
-        """
-        Filters out all documents which have more or less than 1 class.
-        Then filters out all classes which have no remaining documents.
-
-        Args:
-            r8 (bool, optional): R8 is constructed by taking only the top 10 (original) classes. Defaults to False.
-            val_size (float, optional): Proportion of training documents to include in the validation set.
-
-        Returns:
-            doc_splits (tuple): Tuple containing 3 List of training, test, and validation documents.
-            unique_classes (List): List of Strings containing the class names sorted in alphabetical order.
-        """
-        # Filter out docs which don't have exactly 1 class
-        data = defaultdict(lambda: {'train': [], 'test': []})
-
-        for doc in reuters.fileids():
-            categories = reuters.categories(doc)
-            if len(categories) == 1:
-                if doc.startswith('training'):
-                    data[categories[0]]['train'].append(doc)
-                if doc.startswith('test'):
-                    data[categories[0]]['test'].append(doc)
-
-        # Filter out classes which have no remaining docs
-        for cat in reuters.categories():
-            if len(data[cat]['train']) < 1 or len(data[cat]['test']) < 1:
-                data.pop(cat, None)
-
-        if r8:
-            # Choose top 10 classes and then select the ones which still remain after filtering
-            popular = sorted(reuters.categories(), key=lambda clz: len(reuters.fileids(clz)), reverse=True)[:10]
-            data = dict([(cls, splits) for (cls, splits) in data.items() if cls in popular])
-
-        # Create splits
-        train_docs = [doc for cls, splits in data.items() for doc in splits['train']]
-        test_docs = [doc for cls, splits in data.items() for doc in splits['test']]
-
-        # Select the validation documents out of the training documents
-        val_size = int(len(train_docs) * val_size)
-        random.shuffle(train_docs)
-        val_docs = train_docs[:val_size]
-        train_docs = train_docs[val_size:]
-
-        # sort the unique classes to ensure constant order
-        unique_classes = sorted(data.keys())
-
-        return (train_docs, test_docs, val_docs), unique_classes
-
-    @staticmethod
     def _prepare_split(docs, classes):
         """
         Extract the text and labels of each documents.
@@ -118,7 +62,7 @@ class Reuters(TextDataset):
             texts.append(text)
             labels.append(classes.index(clz))
 
-        return texts, labels
+        return texts[:100], labels[:100]
     
     def get_collate_fn(self):
         """
@@ -140,7 +84,7 @@ class Reuters(TextDataset):
         return len(self.labels)
 
 
-class R52(Reuters):
+class R52Text(ReutersText):
     """
     Wrapper for the R52 dataset.
     """
@@ -153,7 +97,7 @@ class R52(Reuters):
         return super().splits(r8=False, *args, **kwargs)
 
 
-class R8(Reuters):
+class R8Text(ReutersText):
     """
     Wrapper for the R8 dataset.
     """
@@ -172,7 +116,7 @@ if __name__ == "__main__":
 
     tokenizer = RobertaTokenizer.from_pretrained("roberta-base")
 
-    train_set, test_set, val_set = R52.splits(tokenizer)
+    train_set, test_set, val_set = R52Text.splits(tokenizer)
     print("train size=", len(train_set))
     print("test size=", len(test_set))
     print("val size=", len(val_set))
